Move deserialization code to NewRelic.Core to avoid dependency
