+++ src/NewRelic.Core/JsonConverters/LambdaPayloads/SnsMessage.cs
@@ -1,10 +1,9 @@
 // Copyright 2020 New Relic, Inc. All rights reserved.
 // SPDX-License-Identifier: Apache-2.0
 
-using System;
 using System.Collections.Generic;
 
-namespace NewRelic.Agent.Extensions.Lambda
+namespace NewRelic.Core.JsonConverters.LambdaPayloads
 {
 
     public class SnsMessage
